Remove RuleOverrideTileBase
