--- Editor/Tiles/RuleOverrideTile/RuleOverrideTileBaseEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0de95f61603644d0b936078ea331196a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
