--- Runtime/Tiles/RuleOverrideTile/IRuleOverrideTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 605f10a6421de4ab78d43e0c7387e636
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
