+++ Runtime/Tiles/RuleOverrideTile/RuleOverrideTile.cs
@@ -11,9 +11,12 @@ namespace UnityEngine.Tilemaps
     [MovedFrom(true, "UnityEngine")]
     [Serializable]
     [CreateAssetMenu(fileName = "New Rule Override Tile", menuName = "Tiles/Rule Override Tile")]
-    public class RuleOverrideTile : RuleOverrideTileBase, IRuleOverrideTile<Sprite, Sprite>, IRuleOverrideTile<GameObject, GameObject>
+    public class RuleOverrideTile : TileBase
     {
 
+        /// <summary>
+        /// A data structure storing the Sprite overriding the original RuleTile Sprite
+        /// </summary>
         [Serializable]
         public class TileSpritePair
         {
@@ -21,6 +24,9 @@ namespace UnityEngine.Tilemaps
             public Sprite m_OverrideSprite;
         }
 
+        /// <summary>
+        /// A data structure storing the GameObject overriding the original RuleTile GameObject
+        /// </summary>
         [Serializable]
         public class TileGameObjectPair
         {
@@ -28,6 +34,10 @@ namespace UnityEngine.Tilemaps
             public GameObject m_OverrideGameObject;
         }
 
+        /// <summary>
+        /// Gets the overriding Sprite of a given Sprite. 
+        /// </summary>
+        /// <param name="original">The original Sprite that is overridden</param>
         public Sprite this[Sprite originalSprite]
         {
             get
@@ -66,6 +76,10 @@ namespace UnityEngine.Tilemaps
             }
         }
 
+        /// <summary>
+        /// Gets the overriding GameObject of a given GameObject. 
+        /// </summary>
+        /// <param name="original">The original GameObject that is overridden</param>
         public GameObject this[GameObject originalGameObject]
         {
             get
@@ -104,11 +118,31 @@ namespace UnityEngine.Tilemaps
             }
         }
 
+        /// <summary>
+        /// The RuleTile to override
+        /// </summary>
+        public RuleTile m_Tile;
+        /// <summary>
+        /// A list of Sprite Overrides
+        /// </summary>
         public List<TileSpritePair> m_Sprites = new List<TileSpritePair>();
+        /// <summary>
+        /// A list of GameObject Overrides
+        /// </summary>
         public List<TileGameObjectPair> m_GameObjects = new List<TileGameObjectPair>();
+
+        /// <summary>
+        /// Returns the Rule Tile for retrieving TileData
+        /// </summary>
+        [HideInInspector] public RuleTile m_InstanceTile;
         [NonSerialized] public int m_MissingSpriteIndex = -1;
         [NonSerialized] public int m_MissingGameObjectIndex = -1;
 
+        /// <summary>
+        /// Applies overrides to this
+        /// </summary>
+        /// <param name="overrides">A list of overrides to apply</param>
+        /// <exception cref="ArgumentNullException">The input overrides list is not valid</exception>
         public void ApplyOverrides(IList<KeyValuePair<Sprite, Sprite>> overrides)
         {
             if (overrides == null)
@@ -118,6 +152,11 @@ namespace UnityEngine.Tilemaps
                 this[overrides[i].Key] = overrides[i].Value;
         }
 
+        /// <summary>
+        /// Applies overrides to this
+        /// </summary>
+        /// <param name="overrides">A list of overrides to apply</param>
+        /// <exception cref="ArgumentNullException">The input overrides list is not valid</exception>
         public void ApplyOverrides(IList<KeyValuePair<GameObject, GameObject>> overrides)
         {
             if (overrides == null)
@@ -127,6 +166,11 @@ namespace UnityEngine.Tilemaps
                 this[overrides[i].Key] = overrides[i].Value;
         }
 
+        /// <summary>
+        /// Gets overrides for this
+        /// </summary>
+        /// <param name="overrides">A list of overrides to fill</param>
+        /// <exception cref="ArgumentNullException">The input overrides list is not valid</exception>
         public void GetOverrides(List<KeyValuePair<Sprite, Sprite>> overrides)
         {
             if (overrides == null)
@@ -157,6 +201,11 @@ namespace UnityEngine.Tilemaps
                 overrides.Add(new KeyValuePair<Sprite, Sprite>(sprite, this[sprite]));
         }
 
+        /// <summary>
+        /// Gets overrides for this
+        /// </summary>
+        /// <param name="overrides">A list of overrides to fill</param>
+        /// <exception cref="ArgumentNullException">The input overrides list is not valid</exception>
         public void GetOverrides(List<KeyValuePair<GameObject, GameObject>> overrides)
         {
             if (overrides == null)
@@ -186,7 +235,7 @@ namespace UnityEngine.Tilemaps
                 overrides.Add(new KeyValuePair<GameObject, GameObject>(gameObject, this[gameObject]));
         }
 
-        public override void Override()
+        public virtual void Override()
         {
             if (!m_Tile || !m_InstanceTile)
                 return;
@@ -215,5 +264,84 @@ namespace UnityEngine.Tilemaps
                 tile.m_TilingRules.Add(instanceRule);
             }
         }
+
+        /// <summary>
+        /// Retrieves any tile animation data from the scripted tile.
+        /// </summary>
+        /// <param name="position">Position of the Tile on the Tilemap.</param>
+        /// <param name="tilemap">The Tilemap the tile is present on.</param>
+        /// <param name="tileAnimationData">Data to run an animation on the tile.</param>
+        /// <returns>Whether the call was successful.</returns>
+        public override bool GetTileAnimationData(Vector3Int position, ITilemap tilemap, ref TileAnimationData tileAnimationData)
+        {
+            if (!m_InstanceTile)
+                return false;
+            return m_InstanceTile.GetTileAnimationData(position, tilemap, ref tileAnimationData);
+        }
+
+        /// <summary>
+        /// Retrieves any tile rendering data from the scripted tile.
+        /// </summary>
+        /// <param name="position">Position of the Tile on the Tilemap.</param>
+        /// <param name="tilemap">The Tilemap the tile is present on.</param>
+        /// <param name="tileData">Data to render the tile.</param>
+        public override void GetTileData(Vector3Int position, ITilemap tilemap, ref TileData tileData)
+        {
+            if (!m_InstanceTile)
+                return;
+            m_InstanceTile.GetTileData(position, tilemap, ref tileData);
+        }
+
+        /// <summary>
+        /// This method is called when the tile is refreshed.
+        /// </summary>
+        /// <param name="location">Position of the Tile on the Tilemap.</param>
+        /// <param name="tileMap">The Tilemap the tile is present on.</param>
+        public override void RefreshTile(Vector3Int position, ITilemap tilemap)
+        {
+            if (!m_InstanceTile)
+                return;
+            m_InstanceTile.RefreshTile(position, tilemap);
+        }
+
+        /// <summary>
+        /// StartUp is called on the first frame of the running Scene.
+        /// </summary>
+        /// <param name="location">Position of the Tile on the Tilemap.</param>
+        /// <param name="tilemap">The Tilemap the tile is present on.</param>
+        /// <param name="instantiateedGameObject">The GameObject instantiated for the Tile.</param>
+        /// <returns>Whether StartUp was successful</returns>
+        public override bool StartUp(Vector3Int position, ITilemap tilemap, GameObject go)
+        {
+            if (!m_InstanceTile)
+                return true;
+            return m_InstanceTile.StartUp(position, tilemap, go);
+        }
+
+        /// <summary>
+        /// Copies a Tiling Rule from a given Tiling Rule
+        /// </summary>
+        /// <param name="from">A Tiling Rule to copy from</param>
+        /// <param name="to">A Tiling Rule to copy to</param>
+        public static void CopyTilingRule(RuleTile.TilingRuleOutput from, RuleTile.TilingRuleOutput to)
+        {
+            to.m_Id = from.m_Id;
+            to.m_Sprites = from.m_Sprites.Clone() as Sprite[];
+            to.m_GameObject = from.m_GameObject;
+            to.m_AnimationSpeed = from.m_AnimationSpeed;
+            to.m_PerlinScale = from.m_PerlinScale;
+            to.m_Output = from.m_Output;
+            to.m_ColliderType = from.m_ColliderType;
+            to.m_RandomTransform = from.m_RandomTransform;
+        }
+        public static void CopyTilingRule(RuleTile.TilingRule from, RuleTile.TilingRule to)
+        {
+            CopyTilingRule(from as RuleTile.TilingRuleOutput, to as RuleTile.TilingRuleOutput);
+
+            to.m_Neighbors = from.m_Neighbors;
+            to.m_NeighborPositions = from.m_NeighborPositions;
+            to.m_RuleTransform = from.m_RuleTransform;
+        }
+
     }
 }
