--- Runtime/Tiles/RuleOverrideTile/RuleOverrideTileBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 511c98a9306a243e1ae834dc36da0163
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
